refactor(ui): split large feature views
